Basic adaptation:
Added a new model both Blue and Red

The Red and Blue player models are now usable and basic mock ups.

Also removed the targetVisible text since the targeting boxes work.

Increased the view range and made the world background (skybox) super large but not move with the player (this was to remove a bug in the direct x drawing rejieme). 

The models have also been added to the model folder.